GCP_api_test/requirement.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
dingtalkrobot==2020.10.27.1721
idna==2.10
importlib-metadata==4.0.1
itchat==1.3.10
lxml==4.6.3
more-itertools==8.7.0
paramiko==2.7.2
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
pypng==0.0.20
PyQRCode==1.2.1
pytest==4.5.0
pytest-base-url==1.4.2
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sshtunnel==0.4.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
